updating values in a table
Here we are inserting and updating values in a table. We are using the same table as in the previous example. The following example shows how to insert and update values in a table.

The intention here is to mirror the values from the spreadsheet and
makes it as much update as possible in the database.

Some times the spreadsheet is the source of truth.