feat: add delete people functionality
this commit adds the delete people functionality to the import command.

the delete people functionality is based on the comparison between the
temporary table and the people table.

the comparison is made by the id, name and surname columns.

if the people exists in the database and not exists in the sheet,
the people is deleted from the database.